refactor(security): move AuthNoCacheFilter to the correct package and update imports in SecurityConfig and SecurityConfigTest
