+++ src/test/java/com/renewsim/backend/auth_service/config/SecurityIntegrationTest.java
@@ -87,5 +87,3 @@ class SecurityIntegrationTest {
                 .andExpect(status().isForbidden());
     }
 }
-
-
